requirements.txt
flask~=2.1.2
requests~=2.27.1
hazelcast-python-client
python-consul